requirements.txt
websockets==15.0.1
matplotlib>=3.8
python-dotenv>=1.0.1
